PBIX FIxing Issues faced on the interview
This Pbix files is a second version of my test. Durin the interview a strange behavior on the function datesytd wasn't reflection the expected behavior. So I went back and made some changes on how i extracted the Date on my Fact tables to establish a relationship with de calendar dimension. The original file i got the date from the "time_hour" column to obtain the date column.
On this version I got the date from the columns year, month and day, so i have only observations from the 2013 year, and the behavior of the function datesytd worked properly.